Created utils file and moved React UI components to a folder
