+++ src/routes/safe/components/AddressBook/components/ImportEntriesModal/index.tsx
@@ -5,9 +5,9 @@ import { Text } from '@gnosis.pm/safe-react-components'
 import { Modal } from 'src/components/Modal'
 import { CSVReader } from 'react-papaparse'
 import { AddressBookEntry } from 'src/logic/addressBook/model/addressBook'
-import { isValidAddress } from 'src/utils/isValidAddress'
 import { checksumAddress } from 'src/utils/checksumAddress'
-import HelpInfo from 'src/routes/safe/components/AddressBook/HelpInfo'
+import HelpInfo from 'src/routes/safe/components/AddressBook/components/HelpInfo'
+import { validateCsvData, validateFile } from '../../utils'
 
 const ImportContainer = styled.div`
   flex-direction: column;
@@ -27,15 +27,6 @@ const InfoContainer = styled.div`
   text-align: center;
   margin-top: 16px;
 `
-
-const WRONG_FILE_EXTENSION_ERROR = 'Only CSV files are allowed'
-const FILE_SIZE_TOO_BIG = 'The size of the file is over 1 MB'
-const FILE_BYTES_LIMIT = 1000000
-const IMPORT_SUPPORTED_FORMATS = [
-  'application/vnd.ms-excel',
-  'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-  'text/csv',
-]
 
 type ImportEntriesModalProps = {
   importEntryModalHandler: (addressList: AddressBookEntry[]) => void
@@ -74,36 +65,6 @@ const ImportEntriesModal = ({ importEntryModalHandler, isOpen, onClose }: Import
     setEntryList(formatedList)
     setImportError('')
     setCsvLoaded(true)
-  }
-
-  const validateFile = (file) => {
-    if (!IMPORT_SUPPORTED_FORMATS.includes(file.type)) {
-      return WRONG_FILE_EXTENSION_ERROR
-    }
-
-    if (file.size >= FILE_BYTES_LIMIT) {
-      return FILE_SIZE_TOO_BIG
-    }
-
-    return
-  }
-
-  const validateCsvData = (data) => {
-    for (let index = 0; index < data.length; index++) {
-      const entry = data[index]
-      if (!entry.data[0] || !entry.data[1] || !entry.data[2]) {
-        return `Invalid amount of columns on row ${index + 1}`
-      }
-      // Verify address properties
-      const address = entry.data[0].toLowerCase()
-      if (!isValidAddress(address)) {
-        return `Invalid address on row ${index + 1}`
-      }
-      if (isNaN(entry.data[2])) {
-        return `Invalid chain id on row ${index + 1}`
-      }
-    }
-    return
   }
 
   const handleOnError = (error) => {
