+++ src/routes/safe/components/Transactions/TxList/TxInfoDetails.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import { fromTokenUnit } from 'src/logic/tokens/utils/humanReadableValue'
 import { ZERO_ADDRESS } from 'src/logic/wallets/ethAddresses'
 import { Erc721Transfer, Transfer } from 'src/logic/safe/store/models/types/gateway.d'
-import { EllipsisTransactionDetails } from 'src/routes/safe/components/AddressBook/EllipsisTransactionDetails'
+import { EllipsisTransactionDetails } from 'src/routes/safe/components/AddressBook/components/EllipsisTransactionDetails'
 import SendModal from 'src/routes/safe/components/Balances/SendModal'
 import { AddressInfo } from './AddressInfo'
 import { InfoDetails } from './InfoDetails'
